requirements/dev.txt
# Application requiremeents
# Main requirements for the application
Django==1.6.2
South==0.8.4
python-social-auth==0.1.22
django-jsonfield==0.9.12
django-tastypie==0.11.0
django-notifications-hq==0.6.0

# Scraping and web parsing
beautifulsoup4==4.3.2
mechanize==0.2.5

# Network CIDR
ipaddress==1.0.6

# Static flies and styling
django-admin-bootstrapped==1.6.2
